requirements.txt
playwright==1.52.0
pyyaml==6.0.2
jieba==0.42.1
httpx==0.28.1
Pillow==9.5.0
moviepy==1.0.3
imagehash==4.3.1
numpy==2.2.5
piper-tts==1.4.1
g2pW==0.1.1
unicode-rbnf==2.4.0
sentence-stream==1.3.0
torch==2.10.0
duckduckgo-search
